[FE] Refactor : 폴더 구조 및 파일 import 문 수정
Title >
* 폴더 구조 및 파일 import 문 수정.

Description >
* utils 폴더 이름 변경에 의한 util 파일 import 문 수정.
